Fix bugs and many compiler warnings
* Promoted size types in CUtlBuffer to int64.
* Fixed bug where CUtlBuffer::GetPtr returns a 64bit pointer that was promoted from a 32bit integral type.
* Use int64/size_t for anything strlen, pointer subtraction, etc..
* Removed invalid structure alignment declarations.

+++ r5dev/public/utility/module.cpp
@@ -69,7 +69,7 @@ CMemory CModule::FindPatternSIMD(const uint8_t* szPattern, const char* szMask, c
 		{
 			if (szMask[i * 16 + j] == 'x')
 			{
-				_bittestandset(reinterpret_cast<LONG*>(&nMasks[i]), j);
+				_bittestandset(reinterpret_cast<LONG*>(&nMasks[i]), static_cast<LONG>(j));
 			}
 		}
 	}
